requirements.txt
openai # for openai api access
mistralai # for mistral ai api access
google-generativeai # for google gemini api access
python-dotenv # for loading environment variables from .env files
requests # for making http requests
markdown # for html generation or markdown processing
tqdm # for displaying progress bars in the console
